moved App.tsx to components/ FreeCodeCamp - Data Visualization Certification - React Projects
